# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/authorization:
    get:
      operationId: getCurrentAdminAuthorization
      tags: [admin]
      summary: Return the current Admin member and effective permissions
      security:
        - OperatorSession: []
      responses:
        "200":
          description: Current member authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthorization"
  /admin/team-members:
    get:
      operationId: listTeamMembers
      tags: [admin]
      summary: List Admin team members
      security:
        - OperatorSession: []
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TeamMember" }
    post:
      operationId: inviteTeamMember
      tags: [admin]
      summary: Invite a team member
      security:
        - OwnerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeamMemberInvite" }
      responses:
        "201":
          description: Invited team member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
  /admin/team-members/{id}:
    patch:
      operationId: updateTeamMember
      tags: [admin]
      summary: Update a team member with optimistic concurrency
      security:
        - OwnerSession: []
      parameters: [{ $ref: "#/components/parameters/OperatorId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeamMemberUpdate" }
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
    delete:
      operationId: removeTeamMember
      tags: [admin]
      summary: Remove a team member
      security:
        - OwnerSession: []
      parameters: [{ $ref: "#/components/parameters/OperatorId" }]
      responses:
        "204": { description: Team member removed }
  /admin/team-members/{id}/suspend:
    post:
      operationId: suspendTeamMember
      tags: [admin]
      summary: Suspend a team member
      security:
        - OwnerSession: []
      parameters: [{ $ref: "#/components/parameters/OperatorId" }]
      responses:
        "200":
          description: Suspended team member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
  /admin/team-members/{id}/reactivate:
    post:
      operationId: reactivateTeamMember
      tags: [admin]
      summary: Reactivate a team member
      security:
        - OwnerSession: []
      parameters: [{ $ref: "#/components/parameters/OperatorId" }]
      responses:
        "200":
          description: Reactivated team member
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
  /admin/team-roles:
    get:
      operationId: listTeamRoles
      tags: [admin]
      summary: List reusable team roles
      security:
        - OperatorSession: []
      responses:
        "200":
          description: Reusable roles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TeamRole" }
    post:
      operationId: createTeamRole
      tags: [admin]
      summary: Create a reusable team role
      security:
        - OwnerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeamRoleInput" }
      responses:
        "201":
          description: Created role
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamRole" }
  /admin/team-roles/{id}:
    patch:
      operationId: updateTeamRole
      tags: [admin]
      summary: Update a reusable team role
      security:
        - OwnerSession: []
      parameters: [{ $ref: "#/components/parameters/OperatorId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeamRoleInput" }
      responses:
        "200":
          description: Updated role
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamRole" }
    delete:
      operationId: deleteTeamRole
      tags: [admin]
      summary: Delete an unused reusable team role
      security:
        - OwnerSession: []
      parameters: [{ $ref: "#/components/parameters/OperatorId" }]
      responses:
        "204": { description: Role deleted }
  /admin/activity:
    get:
      operationId: listAdminActivity
      tags: [admin]
      summary: List safe Admin activity events
      security:
        - OperatorSession: []
      parameters:
        - name: cursor
          in: query
          schema: { type: string, maxLength: 512 }
        - name: section
          in: query
          schema: { type: string, maxLength: 80 }
        - name: permissionKey
          in: query
          schema: { type: string, maxLength: 120 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Keyset-paged activity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminActivityPage" }
...
components:
  parameters:
...
    OperatorId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    CustomerSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk customer session. Order ownership is derived from this authenticated session on the server.
    OperatorSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk operator session. The authenticated user's verified email must match an active server-side operator account.
    OwnerSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk owner session. The authenticated user must match the single active owner account in the server-side operator registry.
  schemas:
...
    PermissionKey:
      type: string
      enum:
        - orders.view
        - orders.details
        - orders.support_tools
        - orders.status
        - orders.confirm_payment
        - orders.complete
        - orders.cancel
        - orders.notes
        - orders.search
        - orders.export
        - orders.assign
        - orders.archive
        - customers.view
        - customers.edit
        - customers.suspend
        - customers.password_reset
        - crypto_assets.view
        - crypto_assets.manage
        - crypto_networks.view
        - crypto_networks.manage
        - crypto_networks.receiving_wallets
        - payment_methods.view
        - payment_methods.manage
        - pricing.view
        - pricing.manage
        - receiving_wallets.view
        - receiving_wallets.manage
        - currencies.view
        - currencies.manage
        - integrations.view
        - integrations.credentials.create
        - integrations.credentials.update
        - integrations.credentials.delete
        - integrations.credentials.rotate
        - integrations.credentials.test
        - blog.view
        - blog.manage
        - languages.view
        - languages.manage
        - social_media.view
        - social_media.manage
        - site_settings.view
        - site_settings.manage
        - statistics.view
        - statistics.export
        - team.members.view
        - team.members.invite
        - team.members.update
        - team.members.suspend
        - team.members.remove
        - team.roles.view
        - team.roles.create
        - team.roles.update
        - team.roles.delete
        - team.permissions.individual
        - team.owner.lifecycle
        - team.activity.view
        - affiliates.view
        - affiliates.manage
    PermissionCatalogEntry:
      type: object
      required: [key, section, label, ownerOnly]
      properties:
        key: { $ref: "#/components/schemas/PermissionKey" }
        section: { type: string }
        label: { type: string }
        ownerOnly: { type: boolean }
    AdminAuthorization:
      type: object
      required: [member, owner, effectivePermissions, catalog]
      properties:
        member: { $ref: "#/components/schemas/TeamMember" }
        owner: { type: boolean }
        effectivePermissions: { type: array, items: { $ref: "#/components/schemas/PermissionKey" } }
        catalog:
          type: array
          items: { $ref: "#/components/schemas/PermissionCatalogEntry" }
    TeamMember:
      type: object
      required: [id, email, name, role, status, customRoleId, permissionAllows, permissionDenies, effectivePermissions, linkedToClerk, authVersion, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, maxLength: 320 }
        name: { type: string }
        role: { type: string, enum: [owner, operator] }
        status: { type: string, enum: [invited, active, suspended, removed] }
        customRoleId: { type: [string, "null"], format: uuid }
        permissionAllows: { type: array, items: { $ref: "#/components/schemas/PermissionKey" } }
        permissionDenies: { type: array, items: { $ref: "#/components/schemas/PermissionKey" } }
        effectivePermissions: { type: array, items: { $ref: "#/components/schemas/PermissionKey" } }
        linkedToClerk: { type: boolean }
        authVersion: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TeamMemberInvite:
      type: object
      required: [name, email]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        email: { type: string, maxLength: 320, pattern: '^[^@\s]+@[^@\s]+\.[^@\s]+$' }
        customRoleId: { type: [string, "null"], format: uuid }
        permissionAllows: { type: array, maxItems: 100, items: { $ref: "#/components/schemas/PermissionKey" } }
        permissionDenies: { type: array, maxItems: 100, items: { $ref: "#/components/schemas/PermissionKey" } }
    TeamMemberUpdate:
      type: object
      required: [expectedAuthVersion]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        status: { type: string, enum: [invited, active, suspended] }
        customRoleId: { type: [string, "null"], format: uuid }
        permissionAllows: { type: array, maxItems: 100, items: { $ref: "#/components/schemas/PermissionKey" } }
        permissionDenies: { type: array, maxItems: 100, items: { $ref: "#/components/schemas/PermissionKey" } }
        expectedAuthVersion: { type: integer, minimum: 1 }
    TeamRole:
      type: object
      required: [id, name, normalizedName, description, permissionKeys, createdAt, updatedAt, usageCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        normalizedName: { type: string }
        description: { type: string }
        permissionKeys: { type: array, items: { $ref: "#/components/schemas/PermissionKey" } }
        usageCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TeamRoleInput:
      type: object
      required: [name, description, permissionKeys]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 500 }
        permissionKeys: { type: array, maxItems: 100, items: { $ref: "#/components/schemas/PermissionKey" } }
    AdminActivityEvent:
      type: object
      required: [id, member, action, section, entityKind, entityId, safeLabel, occurredAt, outcome]
      properties:
        id: { type: string, format: uuid }
        member: { type: string }
        action: { type: string }
        section: { type: string }
        entityKind: { type: [string, "null"] }
        entityId: { type: [string, "null"] }
        safeLabel: { type: [string, "null"] }
        occurredAt: { type: string, format: date-time }
        outcome: { type: string }
    AdminActivityPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AdminActivityEvent" }
        nextCursor: { type: [string, "null"] }
...